refactor: remove AI routine generation — agent chat replaces it
Deletes the standalone draft/confirm wizard backend: domain/ai (services,
validator, generators, 12 DTOs), AiRoutineController, AiGenerationException
+ handler + 3 orphan ErrorKeys, the 10/hour AI rate-limit bucket, the
routine-generation.st prompt, and 9 feature test files (incl.
RateLimitFilterAiTest).

AiIconCatalog (+ test) MOVES to domain/aiAgent — the agent's tools and
system prompt still use it.

Config: spring.ai.model.chat=deepseek makes DeepSeek the only auto-configured
chat model (the chain builds free providers manually); a noop
spring.ai.openai.api-key stays because the starter's other modality
autoconfigs (audio/embedding) demand a credential at boot. AI_API_KEY /
AI_ROUTINE_MODEL / AI_ROUTINE_ENABLED dropped from envExample.

Suite: 484 tests green. Includes the removal spec + plan under
docs/superpowers/.
