+++ CLAUDE.md
@@ -51,7 +51,7 @@ Base package: `beyou.beyouapp.backend`
 
 **RefreshUiDTO**: After mutations (check/uncheck habits in routines, complete goals, etc.), the backend returns a `RefreshUiDTO` containing updated XP/level data for user, categories, habits, and the checked item — so the frontend can update all affected UI in one response. (Note: untyped on the frontend side — silent break risk on field renames.)
 
-**LLM fallback chain** (2026-07-17, branch `llm_fallback`): the agent chat's `ChatModel` is a `FallbackChatModel` (`domain/aiAgent/llm/`, `@Primary`) walking `ai.llm-chain.order` (default mistral→gemini→openrouter→deepseek — openrouter is the free daily buffer (50 req/day, 1000/day after a one-time $10 credit) before paid DeepSeek; groq/cerebras stay configured but out of the default order — groq's 6k-TPM free tier is smaller than the agent prompt and cerebras free tier 402s, re-add via `LLM_CHAIN_ORDER` if their tiers improve). Free providers are OpenAI-compatible manual `OpenAiChatModel`s built via `OpenAiSetup.setupSyncClient/setupAsyncClient` (both clients required: call() is sync, stream() is async) — NOT Spring beans, so `SpringAiRoutineDraftGenerator`'s auto-configured `OpenAiChatModel` injection stays unambiguous. Providers without an API key are skipped at boot (dev/e2e run DeepSeek-only). Fallback fires ONLY before a delegate's first emission — tools never re-run (Spring AI 2.0 runs the tool loop in the ChatClient's `ToolCallingAdvisor`, above the model). CRITICAL options contract: `FallbackChatModel.getOptions()` must return a portable `ToolCallingChatOptions` (the ChatClient seeds prompt options from the injected model's `getOptions()` — the 2.0 method, NOT the deprecated `getDefaultOptions()` — and the `ToolCallingAdvisor` only engages on that type), and `adaptFor()` retypes prompt options per delegate because every provider model hard-casts prompt options to its native class (`OpenAiChatOptions`, `DeepSeekChatOptions`). 429 → 300s cooldown, other errors → 30s; the last chain link always plays even in cooldown. Metrics: `beyou.ai.llm.calls{provider,outcome}`, `beyou.ai.llm.fallback{from,reason}`, `beyou.ai.llm.exhausted`; Grafana row "Fallback chain" in `Beyou-dev-env` `beyou-ai-agent.json`. Routine generation stays pinned to OpenAI, outside the chain.
+**LLM fallback chain** (2026-07-17, branch `llm_fallback`): the agent chat's `ChatModel` is a `FallbackChatModel` (`domain/aiAgent/llm/`, `@Primary`) walking `ai.llm-chain.order` (default mistral→gemini→openrouter→deepseek — openrouter is the free daily buffer (50 req/day, 1000/day after a one-time $10 credit) before paid DeepSeek; groq/cerebras stay configured but out of the default order — groq's 6k-TPM free tier is smaller than the agent prompt and cerebras free tier 402s, re-add via `LLM_CHAIN_ORDER` if their tiers improve). Free providers are OpenAI-compatible manual `OpenAiChatModel`s built via `OpenAiSetup.setupSyncClient/setupAsyncClient` (both clients required: call() is sync, stream() is async) — NOT Spring beans; DeepSeek is the only auto-configured chat model (`spring.ai.model.chat: deepseek`; a noop `spring.ai.openai.api-key` keeps the starter's other modality autoconfigs booting). Providers without an API key are skipped at boot (dev/e2e run DeepSeek-only). Fallback fires ONLY before a delegate's first emission — tools never re-run (Spring AI 2.0 runs the tool loop in the ChatClient's `ToolCallingAdvisor`, above the model). CRITICAL options contract: `FallbackChatModel.getOptions()` must return a portable `ToolCallingChatOptions` (the ChatClient seeds prompt options from the injected model's `getOptions()` — the 2.0 method, NOT the deprecated `getDefaultOptions()` — and the `ToolCallingAdvisor` only engages on that type), and `adaptFor()` retypes prompt options per delegate because every provider model hard-casts prompt options to its native class (`OpenAiChatOptions`, `DeepSeekChatOptions`). 429 → 300s cooldown, other errors → 30s; the last chain link always plays even in cooldown. Metrics: `beyou.ai.llm.calls{provider,outcome}`, `beyou.ai.llm.fallback{from,reason}`, `beyou.ai.llm.exhausted`; Grafana row "Fallback chain" in `Beyou-dev-env` `beyou-ai-agent.json`. The standalone AI routine generation feature was REMOVED 2026-07-19 (`domain/ai/` deleted, `AiIconCatalog` moved to `domain/aiAgent/`) — the agent chat covers routine creation via tools.
 
 **Authentication flow**: JWT access token (15min, returned in the `X-Access-Token` response header) + refresh token (15 days, `httpOnly` cookie). `SecurityFilter` validates JWT on every request except public endpoints (`/auth/login`, `/auth/register`, `/auth/refresh`, `/auth/google`, `/auth/logout`, `/auth/forgot-password`, `/auth/reset-password/**`, `/auth/verify-email`, `/docs/**` except `/docs/admin/**`). `AuthenticatedUser` component extracts the current user from `SecurityContextHolder`. Profile `e2e` auto-verifies emails on registration (`e2e.auto-verify-email: true` in `UserService`).
