+++ envExample
@@ -32,11 +32,6 @@ MAIL_FROM=
 PASSWORD_RESET_TTL_MINUTES=30
 PASSWORD_RESET_COOLDOWN_MINUTES=5
 
-# AI (routine generation)
-AI_API_KEY=
-AI_ROUTINE_MODEL=gpt-5-mini
-AI_ROUTINE_ENABLED=true
-
 LLM_CHAIN_ORDER=mistral,gemini,openrouter,deepseek
 MISTRAL_API_KEY=
 GEMINI_API_KEY=
