--- src/main/java/beyou/beyouapp/backend/domain/ai/CannedRoutineDraftGenerator.java
@@ -1,49 +1,0 @@
-package beyou.beyouapp.backend.domain.ai;
-
-import java.util.List;
-import java.util.Set;
-
-import org.springframework.context.annotation.Profile;
-import org.springframework.stereotype.Component;
-
-import beyou.beyouapp.backend.domain.ai.dto.DraftHabitItemDTO;
-import beyou.beyouapp.backend.domain.ai.dto.DraftNewCategoryDTO;
-import beyou.beyouapp.backend.domain.ai.dto.DraftNewHabitDTO;
-import beyou.beyouapp.backend.domain.ai.dto.DraftNewTaskDTO;
-import beyou.beyouapp.backend.domain.ai.dto.DraftSectionDTO;
-import beyou.beyouapp.backend.domain.ai.dto.DraftTaskItemDTO;
-import beyou.beyouapp.backend.domain.ai.dto.RoutineDraftDTO;
-import beyou.beyouapp.backend.domain.routine.schedule.WeekDay;
-
-/**
- * Deterministic draft for the e2e profile: Playwright asserts these exact
- * names (see Beyou-e2e-tests/tests/ai-routine.spec.ts). No API key, no cost,
- * no flake. Keep names in sync with that spec.
- */
-@Component
-@Profile("e2e")
-public class CannedRoutineDraftGenerator implements RoutineDraftGenerator {
-
-    @Override
-    public RoutineDraftDTO generate(GenerationContext ctx) {
-        return new RoutineDraftDTO(
-                "AI Morning Routine",
-                "lucide:sun",
-                List.of(new DraftNewCategoryDTO("new-1", "Wellness", "lucide:leaf", "Generated by AI e2e stub")),
-                List.of(new DraftSectionDTO("Morning", "lucide:sun", "06:00", "09:00",
-                        List.of(
-                                new DraftHabitItemDTO(null,
-                                        new DraftNewHabitDTO("Drink water", "Right after waking up",
-                                                "Stay hydrated!", "lucide:glass-water", 3, 1, List.of("new-1")),
-                                        "06:00", "06:10"),
-                                new DraftHabitItemDTO(null,
-                                        new DraftNewHabitDTO("Stretch", null, null,
-                                                "lucide:brain", 2, 2, List.of("new-1")),
-                                        "06:10", "06:20")),
-                        List.of(new DraftTaskItemDTO(null,
-                                new DraftNewTaskDTO("Prepare breakfast", null,
-                                        "lucide:egg", 2, 1, List.of("new-1"), false),
-                                "06:20", "06:40")))),
-                Set.of(WeekDay.Monday, WeekDay.Tuesday, WeekDay.Wednesday, WeekDay.Thursday, WeekDay.Friday));
-    }
-}
