--- src/main/java/beyou/beyouapp/backend/domain/ai/dto/GenerateRoutineResponseDTO.java
@@ -1,4 +1,0 @@
-package beyou.beyouapp.backend.domain.ai.dto;
-
-public record GenerateRoutineResponseDTO(RoutineDraftDTO draft) {
-}
