+++ src/main/java/beyou/beyouapp/backend/exceptions/ErrorKey.java
@@ -62,9 +62,6 @@ public enum ErrorKey {
     REFRESH_TOKEN_EXPIRED,
     REFRESH_TOKEN_INVALID,
     GOOGLE_OAUTH_FAILED,
-    AI_GENERATION_FAILED,
-    AI_RESPONSE_INVALID,
-    AI_DRAFT_INVALID,
     RATE_LIMIT_EXCEEDED,
     PHOTO_UPLOAD_NO_FILE,
     PHOTO_UPLOAD_INVALID_TYPE,
